skills/create-image/scripts/pyproject.toml
[project]
name = "image-gen"
version = "0.1.0"
description = "Quick CLI for generating images with Gemini Nano Banana Pro."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.64.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
